Fixing Locality
I was able to resolve the last of the local-dependent issues. Therefore, this repository should now be able to run on any system without any direct editing of the codebase. I provided several strategically placed read me files to help in the proper setup of necessary dependencies. They are in src/coreference and dataset/