Classifier for Hotel or Car
I updated the classifier again. I modified the approach to better suit our problem at hand. Specifically, I made the files and directories that the classifier first encounters at the specified path the categories, and any subfiles and subdirectories belong to their top-most category. This was a better approach to the hotel vs car problem than treating each file to load as a separate category.

I updated the SentimentRanker to use the updated Classifier. I also had to clean the hotel files, which accounts for most of the changed files in this update.