Added a new word grouping variant
I made a new variant to the word grouping idea. This new method limits the number of terms in a group by requiring members of the group to have at least a certain percent of the terms in the group. This creates more word redundancy between groups, but prevents groups from running away in size.

I also removed the obsolete test-reviews.txt file.