Set up Alterations
I was trying to set up the project again, and I think that these changes were necessary to remove errors.

I also should mention that Eclipse was not able to find the Stanford POS Tagger, perhaps because of its repository location. Thus I just downloaded the .jar and added it to the build path. You can find the jar at the location given in the comment of the pom.xml.